Added a test for invalid or illegal XML characters. This relates to issue #12 "Malformed HTML cause parsing exception" - the org.w3c.dom.DOMException which had been thrown by the parser is now wrapped in a org.owasp.validator.html.ScanException.
This test uses a selection of strings generated randomly and known
to cause the parser to fail. Additionally, a test case based on the
examples given in the trouble ticket is included.

Note that (strangely) at least one string which caused the parser to 
fail under JDK 1.5 and 1.6 (on OS 10.5.3) did *not* cause the parser
to fail under JDK 1.4. I don't know what causes this inconsistency, but
the test cases included behave consistently across all three runtimes.